# data_aggregator_service/data-aggregator-service-openapi.yml
openapi: 3.0.3
info:
  title: Data Aggregator Service API
  description: |
    Микросервис для агрегации данных о перемещениях объектов, генерации отчетов и аналитики.
    Периодически собирает данные от positioning-service, агрегирует их и предоставляет аналитические отчеты.
  version: 1.0.0
  contact:
    name: Команда разработки
    email: analytics@rtls.example.com
servers:
  - url: http://localhost:8085/api/v1
    description: Локальный сервер разработки
  - url: https://analytics.rtls.example.com/api/v1
    description: Продуктивный сервер
...
paths:
  /aggregation/trigger:
    post:
      tags:
        - Aggregation
      summary: Запуск процесса агрегации данных
      description: Ручной запуск процесса агрегации данных за указанный период
      operationId: triggerAggregation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                start_time:
                  type: string
                  format: date-time
                  description: Начало периода для агрегации
                end_time:
                  type: string
                  format: date-time
                  description: Конец периода для агрегации
                force:
                  type: boolean
                  default: false
                  description: Принудительная агрегация, даже если данные уже существуют
      responses:
        '202':
          description: Процесс агрегации запущен
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    description: Идентификатор задачи агрегации
                  status:
                    type: string
                    enum: [queued, started]
        '400':
          description: Некорректные параметры запроса
...